% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{is.boosted}
\alias{is.boosted}
\title{Is the Object a Boosted Model}
\usage{
is.boosted(x)
}
\arguments{
\item{x}{any \code{R} object}
}
\value{
\code{TRUE} if the object is a boosted model
}
\description{
Test the inheritance of an object
}

