% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{predict.boosted}
\alias{predict.boosted}
\title{Model Predictions}
\usage{
\method{predict}{boosted}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{a \code{boosted} model returned from \link{fitBoosted}}

\item{newdata}{the new independent variables to use for prediction. This should be a data frame.}

\item{...}{additional arguments affecting the predictions produced (ignored).}
}
\value{
\code{predict.boosted} produces a numeric vector with the predicted classes from the \code{boosted} model.
}
\description{
Apply a fitted \code{boosted} model to newdata to form predictions. If no \code{newdata} is included, returned the fitted values of the model.
}

